backend/pyproject.toml
[project]
name = "dvn"
dynamic = ["version"]                                              # let setuptools-scm handle versioning
description = "NLP Food query matcher"
readme = "README.md"
requires-python = "==3.11"

dependencies = [
    "openai",
    "python-dotenv",
    "requests",
    "pandas",
    "bm25s", # https://github.com/xhluca/bm25s
    "PyStemmer", # useful for bm25s
    "jax[cpu]", # bm25 docs say this cna be installed for a speed boost
    "autofaiss>=2.12.0",
    "fsspec>=2024.10.0",
    "numpy"
]

[tool.mypy]
python_version = "3.11"
strict = true

[build-system]
requires = ["setuptools", "wheel", "setuptools-scm[toml]"]
build-backend = "setuptools.build_meta"

# ⬇⬇⬇  THIS tells setuptools what to install  ⬇⬇⬇
[tool.setuptools]
package-dir = { "" = "src" } # “top‑level imports come from src/”

[tool.setuptools_scm]
fallback_version = "0.2.0"

[tool.setuptools.packages.find]
where = ["src"]         # search only there
include = ["prosus*","server*"] # pick up both packages

[tool.pyrefly]
python-version = ">=3.11,<3.12"
